refactor(konnector-libs): Mock global model
