+++ packages/cozy-konnector-libs/src/libs/categorization/__mocks__/set_label_cat.json
